+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Audit function failure due to missing sort methods.
+- More robust error handling for data source historic prices.
 
 ## Version [0.1.3] Beta (2019-08-14)
 ### Fixed